# server/docs/filehandler-openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "File Sharing API",
    "description" : "API for File Sharing Application - Upload, Download, Share files",
    "contact" : {
      "name" : "File Sharing Team",
      "email" : "support@filesharing.com"
    },
    "license" : {
      "name" : "MIT License",
      "url" : "https://opensource.org/licenses/MIT"
    },
    "version" : "1.0.0"
  },
  "servers" : [ {
    "url" : "http://localhost:5000",
    "description" : "Local Development Server"
  }, {
    "url" : "https://api.filesharing.com",
    "description" : "Production Server"
  } ],
  "security" : [ {
    "Bearer Authentication" : [ ]
  } ],
  "paths" : {
...
    "/api/project/{projectId}/visibility" : {
      "post" : {
        "tags" : [ "project-controller" ],
        "operationId" : "updateProjectVisibility",
        "parameters" : [ {
          "name" : "projectId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ProjectVisibilityUpdateDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CommonResponseProjectEntity"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized - Invalid or missing JWT token"
          },
          "403" : {
            "description" : "Forbidden - Access denied"
          },
          "500" : {
            "description" : "Internal Server Error"
          }
        }
      }
    },
    "/api/project/{projectId}/status" : {
      "post" : {
        "tags" : [ "project-controller" ],
        "operationId" : "updateProjectStatus",
        "parameters" : [ {
          "name" : "projectId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ProjectStatusUpdateDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CommonResponseProjectEntity"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized - Invalid or missing JWT token"
          },
          "403" : {
            "description" : "Forbidden - Access denied"
          },
          "500" : {
            "description" : "Internal Server Error"
          }
        }
      }
    },
    "/api/project/{projectId}/share-token/refresh" : {
      "post" : {
        "tags" : [ "project-controller" ],
        "operationId" : "refreshShareToken",
        "parameters" : [ {
          "name" : "projectId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ShareTokenCreateDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CommonResponseShareTokenCreateResponseDTO"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized - Invalid or missing JWT token"
          },
          "403" : {
            "description" : "Forbidden - Access denied"
          },
          "500" : {
            "description" : "Internal Server Error"
          }
        }
      }
    },
    "/api/project/{projectId}/leave" : {
      "post" : {
        "tags" : [ "project-controller" ],
        "operationId" : "leaveProject",
        "parameters" : [ {
          "name" : "projectId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CommonResponseProjectEntity"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized - Invalid or missing JWT token"
          },
          "403" : {
            "description" : "Forbidden - Access denied"
          },
          "500" : {
            "description" : "Internal Server Error"
          }
        }
      }
    },
    "/api/project/{projectId}/addCollaborator" : {
      "post" : {
        "tags" : [ "project-controller" ],
        "operationId" : "addCollaboratorToProject",
        "parameters" : [ {
          "name" : "projectId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ProjectCollaboratorDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CommonResponseProjectEntity"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized - Invalid or missing JWT token"
          },
          "403" : {
            "description" : "Forbidden - Access denied"
          },
          "500" : {
            "description" : "Internal Server Error"
          }
        }
      }
    },
    "/api/project/update-detail" : {
      "post" : {
        "tags" : [ "project-controller" ],
        "operationId" : "updateProjectDetail",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ProjectCreateUpdateDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CommonResponseProjectEntity"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized - Invalid or missing JWT token"
          },
          "403" : {
            "description" : "Forbidden - Access denied"
          },
          "500" : {
            "description" : "Internal Server Error"
          }
        }
      }
    },
    "/api/project/restore/{projectId}" : {
      "post" : {
        "tags" : [ "project-controller" ],
        "operationId" : "restoreProject",
        "parameters" : [ {
          "name" : "projectId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CommonResponseProjectEntity"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized - Invalid or missing JWT token"
          },
          "403" : {
            "description" : "Forbidden - Access denied"
          },
          "500" : {
            "description" : "Internal Server Error"
          }
        }
      }
    },
    "/api/project/remove-collaborator/{projectId}/{collaboratorId}" : {
      "post" : {
        "tags" : [ "project-controller" ],
        "operationId" : "removeCollaboratorFromProject",
        "parameters" : [ {
          "name" : "projectId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "collaboratorId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CommonResponseProjectEntity"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized - Invalid or missing JWT token"
          },
          "403" : {
            "description" : "Forbidden - Access denied"
          },
          "500" : {
            "description" : "Internal Server Error"
          }
        }
      }
    },
    "/api/project/move-to-project/{projectId}" : {
      "post" : {
        "tags" : [ "project-controller" ],
        "operationId" : "moveToProject",
        "parameters" : [ {
          "name" : "projectId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CommonResponseProjectEntity"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized - Invalid or missing JWT token"
          },
          "403" : {
            "description" : "Forbidden - Access denied"
          },
          "500" : {
            "description" : "Internal Server Error"
          }
        }
      }
    },
    "/api/project/get-page" : {
      "post" : {
        "tags" : [ "project-controller" ],
        "operationId" : "getProjectPage",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PageRequestDtoProjectFilterDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CommonResponsePageResultProjectEntity"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized - Invalid or missing JWT token"
          },
          "403" : {
            "description" : "Forbidden - Access denied"
          },
          "500" : {
            "description" : "Internal Server Error"
          }
        }
      }
    },
    "/api/project/create-share-token" : {
      "post" : {
        "tags" : [ "project-controller" ],
        "operationId" : "createShareToken",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ShareTokenCreateDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CommonResponseShareTokenCreateResponseDTO"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized - Invalid or missing JWT token"
          },
          "403" : {
            "description" : "Forbidden - Access denied"
          },
          "500" : {
            "description" : "Internal Server Error"
          }
        }
      }
    },
    "/api/project/create-new" : {
      "post" : {
        "tags" : [ "project-controller" ],
        "operationId" : "createNewProject",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ProjectCreateUpdateDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CommonResponseProjectEntity"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized - Invalid or missing JWT token"
          },
          "403" : {
            "description" : "Forbidden - Access denied"
          },
          "500" : {
            "description" : "Internal Server Error"
          }
        }
      }
    },
    "/api/project/collaborators/changePermission" : {
      "post" : {
        "tags" : [ "project-controller" ],
        "operationId" : "changeCollaboratorPermission",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ProjectCollaboratorDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CommonResponseProjectEntity"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized - Invalid or missing JWT token"
          },
          "403" : {
            "description" : "Forbidden - Access denied"
          },
          "500" : {
            "description" : "Internal Server Error"
          }
        }
      }
    },
    "/api/project/check-project" : {
      "post" : {
        "tags" : [ "project-controller" ],
        "operationId" : "checkProject",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ProjectCheckInputDTO"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CommonResponseProjectCheckResponseDTO"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized - Invalid or missing JWT token"
          },
          "403" : {
            "description" : "Forbidden - Access denied"
          },
          "500" : {
            "description" : "Internal Server Error"
          }
        }
      }
    },
    "/api/project/archive/{projectId}" : {
      "post" : {
        "tags" : [ "project-controller" ],
        "operationId" : "archiveProject",
        "parameters" : [ {
          "name" : "projectId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CommonResponseString"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized - Invalid or missing JWT token"
          },
          "403" : {
            "description" : "Forbidden - Access denied"
          },
          "500" : {
            "description" : "Internal Server Error"
          }
        }
      }
    },
...
    "/api/project/{projectId}/stats" : {
      "get" : {
        "tags" : [ "project-controller" ],
        "operationId" : "getProjectStats",
        "parameters" : [ {
          "name" : "projectId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CommonResponseProjectStats"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized - Invalid or missing JWT token"
          },
          "403" : {
            "description" : "Forbidden - Access denied"
          },
          "500" : {
            "description" : "Internal Server Error"
          }
        }
      }
    },
    "/api/project/{projectId}/members" : {
      "get" : {
        "tags" : [ "project-controller" ],
        "operationId" : "getProjectMembers",
        "parameters" : [ {
          "name" : "projectId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "searchQuery",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "reviewSessionId",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CommonResponseListProjectMemberDTO"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized - Invalid or missing JWT token"
          },
          "403" : {
            "description" : "Forbidden - Access denied"
          },
          "500" : {
            "description" : "Internal Server Error"
          }
        }
      }
    },
    "/api/project/{projectId}/getCollaborators" : {
      "get" : {
        "tags" : [ "project-controller" ],
        "operationId" : "getProjectCollaborators",
        "parameters" : [ {
          "name" : "projectId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CommonResponseListProjectCollaborator"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized - Invalid or missing JWT token"
          },
          "403" : {
            "description" : "Forbidden - Access denied"
          },
          "500" : {
            "description" : "Internal Server Error"
          }
        }
      }
    },
    "/api/project/{projectId}/audit-log" : {
      "get" : {
        "tags" : [ "project-controller" ],
        "operationId" : "getProjectAuditLog",
        "parameters" : [ {
          "name" : "projectId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "dto",
          "in" : "query",
          "required" : true,
          "schema" : {
            "$ref" : "#/components/schemas/PageRequestDtoAuditLogFilterDTO"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CommonResponsePageResultAuditLogEntity"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized - Invalid or missing JWT token"
          },
          "403" : {
            "description" : "Forbidden - Access denied"
          },
          "500" : {
            "description" : "Internal Server Error"
          }
        }
      }
    },
    "/api/project/share-token/{shareToken}" : {
      "get" : {
        "tags" : [ "project-controller" ],
        "operationId" : "getShareTokenInfo",
        "parameters" : [ {
          "name" : "shareToken",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CommonResponseShareTokenInfoDTO"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized - Invalid or missing JWT token"
          },
          "403" : {
            "description" : "Forbidden - Access denied"
          },
          "500" : {
            "description" : "Internal Server Error"
          }
        }
      }
    },
    "/api/project/join-project/{shareToken}" : {
      "get" : {
        "tags" : [ "project-controller" ],
        "operationId" : "joinProject",
        "parameters" : [ {
          "name" : "shareToken",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CommonResponseProjectEntity"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized - Invalid or missing JWT token"
          },
          "403" : {
            "description" : "Forbidden - Access denied"
          },
          "500" : {
            "description" : "Internal Server Error"
          }
        }
      }
    },
    "/api/project/get-by-id/{projectId}" : {
      "get" : {
        "tags" : [ "project-controller" ],
        "operationId" : "getProjectById",
        "parameters" : [ {
          "name" : "projectId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CommonResponseProjectEntity"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized - Invalid or missing JWT token"
          },
          "403" : {
            "description" : "Forbidden - Access denied"
          },
          "500" : {
            "description" : "Internal Server Error"
          }
        }
      }
    },
...
    "/api/project/{projectId}/delete-share-token" : {
      "delete" : {
        "tags" : [ "project-controller" ],
        "operationId" : "revokeShareToken",
        "parameters" : [ {
          "name" : "projectId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CommonResponseString"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized - Invalid or missing JWT token"
          },
          "403" : {
            "description" : "Forbidden - Access denied"
          },
          "500" : {
            "description" : "Internal Server Error"
          }
        }
      }
    },
    "/api/project/delete/{projectId}" : {
      "delete" : {
        "tags" : [ "project-controller" ],
        "operationId" : "deleteProject",
        "parameters" : [ {
          "name" : "projectId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CommonResponseString"
                }
              }
            }
          },
          "401" : {
            "description" : "Unauthorized - Invalid or missing JWT token"
          },
          "403" : {
            "description" : "Forbidden - Access denied"
          },
          "500" : {
            "description" : "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "GrantedVisibility" : {
        "type" : "string",
        "enum" : [ "PUBLIC", "PRIVATE" ]
      },
      "ProjectVisibilityUpdateDTO" : {
        "type" : "object",
        "properties" : {
          "visibility" : {
            "$ref" : "#/components/schemas/GrantedVisibility"
          }
        }
      },
      "CommonResponseProjectEntity" : {
        "type" : "object",
        "properties" : {
          "isSuccessful" : {
            "type" : "boolean"
          },
          "data" : {
            "$ref" : "#/components/schemas/ProjectEntity"
          },
          "code" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          }
        }
      },
...
      "ProjectStatusUpdateDTO" : {
        "type" : "object",
        "properties" : {
          "status" : {
            "$ref" : "#/components/schemas/ProjectStatus"
          }
        }
      },
      "ShareTokenCreateDTO" : {
        "type" : "object",
        "properties" : {
          "rangeTime" : {
            "$ref" : "#/components/schemas/ShareTokenTime"
          },
          "expireDate" : {
            "type" : "string",
            "format" : "date-time"
          },
          "projectId" : {
            "type" : "string"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "Bearer Authentication" : {
        "type" : "http",
        "description" : "Enter JWT token",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}